backend-streetwear/BackendReact/requirements.txt
asgiref==3.7.2
channels==4.0.0x
coverage==7.2.7
Django==4.2.4
django-base64field==1.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.7.0
factory-boy==3.3.0
Faker==19.3.0
filetype==1.2.0
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
